apps/e2e-tests/requirements.txt
# Testing Framework
pytest==8.0.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-timeout==2.2.0
pytest-xdist==3.5.0

# HTTP Testing
httpx==0.26.0
requests==2.31.0

# WebSocket Testing
websockets==12.0

# Assertions & Mocking
pytest-mock==3.12.0
faker==22.0.0

# Performance Testing
locust==2.20.0

# Reporting
pytest-html==4.1.1
allure-pytest==2.13.2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
